chore(compiler): remove circular dependency in bundle
Two changes were necessary to remove a circular dependency relationship
between `bundle-output.ts` and `dev-module.ts`:

- two constants are moved to a new file, `constants.ts`
- functionality relating to resolving node module IDs is moved to a new
  file (`dev-node-module-resolve.ts`)
