+++ src/compiler/bundle/constants.ts
@@ -1,0 +1,3 @@
+export const DEV_MODULE_CACHE_BUSTER = 0;
+
+export const DEV_MODULE_DIR = `~dev-module`;
